backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
pymongo==3.13.0
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.0.1
Werkzeug==3.0.1
